Finally fixed windows compat and fix conversion warnings

+++ include/nn/layers_impl/activation_layer.hpp
@@ -21,15 +21,15 @@ namespace tnn {
 template <typename T = float> class ActivationLayer : public StatelessLayer<T> {
 private:
   std::unique_ptr<ActivationFunction<T>> activation_;
-  std::unordered_map<int, Tensor<T>> micro_batch_inputs_;
+  std::unordered_map<size_t, Tensor<T>> micro_batch_inputs_;
 
 public:
   explicit ActivationLayer(std::unique_ptr<ActivationFunction<T>> activation,
                            const std::string &name = "activation");
 
-  Tensor<T> forward(const Tensor<T> &input, int micro_batch_id = 0) override;
+  Tensor<T> forward(const Tensor<T> &input, size_t micro_batch_id = 0) override;
   Tensor<T> backward(const Tensor<T> &grad_output,
-                     int micro_batch_id = 0) override;
+                     size_t micro_batch_id = 0) override;
 
   std::string type() const override;
   LayerConfig get_config() const override;
